move TaskPackageDropbox to its own file
